file-service/requirements.txt
beanie==1.29.0
chromadb==1.0.8
dotenv==0.9.9
fastapi==0.115.9
fastapi-cli==0.0.7
motor==3.7.0
numpy==2.2.5
ollama==0.4.8
pymongo==4.12.1
PyMuPDF==1.25.5
pypdf==3.17.0
python-docx==1.1.2
python-dotenv==1.1.0
sentence-transformers==4.1.0
sse-starlette==2.3.3
starlette==0.45.3
uvicorn==0.34.2
